Sync

+++ Money app/app/src/main/java/com/example/moneymanager/RecyclerViewAdapter.java
@@ -15,14 +15,29 @@ import android.widget.TextView;
 import com.example.moneymanager.DAO.AppDatabase;
 import com.example.moneymanager.DAO.Category;
 import com.example.moneymanager.DAO.ExpensesAndIncomes;
+import com.example.moneymanager.DAO.TimeStamp;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.firestore.FirebaseFirestore;
 
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Locale;
+import java.util.Map;
 
 public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapter.MyViewHolder> {
 
-    Context context;
-    List<Category> data;
-    AppDatabase db;
+    private Context context;
+    private List<Category> data;
+    private AppDatabase db;
+    private FirebaseUser currentUser;
+    private FirebaseFirestore fireDB;
+    private FirebaseAuth auth;
+
+    private static final String TIME_FORMAT = "dd/MM/yyy HH:mm:ss";
+    private SimpleDateFormat timeFormat = new SimpleDateFormat(TIME_FORMAT, Locale.getDefault());
 
     public RecyclerViewAdapter(Context context, List<Category> data) {
         this.context = context;
@@ -100,13 +115,58 @@ public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapte
                             @Override
                             public void run() {
                                 db = AppDatabase.getInstance(context);
+                                auth = FirebaseAuth.getInstance();
+                                fireDB = FirebaseFirestore.getInstance();
                                 Category category = db.categoryDAO().findById(data.get(position).getId());
                                 List<ExpensesAndIncomes> toDelete = db.expensesAndIncomeDAO().getExpensesForOneCategory(category.getId());
-                                for (int i = 0; i < toDelete.size(); i++) {
-                                    ExpensesAndIncomes item = db.expensesAndIncomeDAO().findById(toDelete.get(i).getId());
-                                    db.expensesAndIncomeDAO().delete(item);
+                                FirebaseUser currentUser = auth.getCurrentUser();
+                                Date date = new Date();
+                                String time = timeFormat.format(date);
+                                Map<String, String> timeMap = new HashMap<>();
+                                timeMap.put("time", time);
+
+                                TimeStamp timeStamp = db.timeStampDAO().getCategoryTime();
+                                if (toDelete.size() != 0) {
+                                    for (int i = 0; i < toDelete.size(); i++) {
+                                        ExpensesAndIncomes item = db.expensesAndIncomeDAO().findById(toDelete.get(i).getId());
+                                        if (category.getType().equals("EXPENSES")) {
+                                            fireDB.collection("Expenses").document(currentUser.getEmail()).collection("Expenses").document(item.getNote())
+                                                    .delete();
+                                            timeStamp.setTimeCategory(time);
+                                            db.timeStampDAO().edit(timeStamp);
+                                            fireDB.collection("Expenses").document(currentUser.getEmail()).collection("Expenses").document("time")
+                                                    .set(timeMap);
+                                        } else {
+                                            fireDB.collection("Expenses").document(currentUser.getEmail()).collection("Incomes").document(category.getName())
+                                                    .delete();
+                                            fireDB.collection("Expenses").document(currentUser.getEmail()).collection("Incomes").document("time")
+                                                    .set(timeMap);
+                                            timeStamp.setTimeCategoryIncome(time);
+                                            db.timeStampDAO().edit(timeStamp);
+                                        }
+                                        db.expensesAndIncomeDAO().delete(item);
+
+                                    }
                                 }
                                 db.categoryDAO().deleteCateogry(category);
+
+
+                                if (category.getType().equals("EXPENSES")) {
+                                    fireDB.collection("Categories").document(currentUser.getEmail()).collection("UserCategoriesExpenses").document(category.getName())
+                                            .delete();
+                                    timeStamp.setTimeCategory(time);
+                                    db.timeStampDAO().edit(timeStamp);
+                                    fireDB.collection("Categories").document(currentUser.getEmail()).collection("UserCategoriesExpenses").document("time")
+                                            .set(timeMap);
+                                } else {
+                                    fireDB.collection("Categories").document(currentUser.getEmail()).collection("UserCategoryIncomes").document(category.getName())
+                                            .delete();
+                                    fireDB.collection("Categories").document(currentUser.getEmail()).collection("UserCategoryIncomes").document("time")
+                                            .set(timeMap);
+                                    timeStamp.setTimeCategoryIncome(time);
+                                    db.timeStampDAO().edit(timeStamp);
+                                }
+
 
                             }
                         });
